Add files via upload
QuizPage.html is the primary starter page or the main interface where the user can choose between the two quizzes, quiz1 and quiz2, that will take them to the pages Quiz1.html and Quiz2.html and finally QuizOneComplete.html and QuizTwoComplete.html respectively. QuizData.json is the JSON file where all the quiz data, questions and answers, are stored in JSON format.